refactor: auto-split test_core_types.py
